+++ Library/Phalcon/Validation/Validator/Iban.php
@@ -21,7 +21,7 @@ namespace Phalcon\Validation\Validator;
 
 use Phalcon\Validation\Validator;
 use Phalcon\Validation;
-use Phalcon\Validation\Message;
+use Phalcon\Messages\Message;
 
 /**
  * Validates IBAN Numbers (International Bank Account Numbers)
