+++ Library/Phalcon/Validation/Validator/MongoId.php
@@ -22,7 +22,7 @@ namespace Phalcon\Validation\Validator;
 use MongoId as Id;
 use Phalcon\Validation;
 use Phalcon\Validation\Validator;
-use Phalcon\Validation\Message;
+use Phalcon\Messages\Message;
 use Phalcon\Validation\Exception as ValidationException;
 
 /**
